backend/requirements.txt
astroid==2.0.4
certifi==2018.10.15
chardet==3.0.4
dataclasses==0.6
Django==2.1.1
django-cors-headers==2.4.0
django-sendgrid-v5==0.6.893
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djongo==1.2.30
future==0.16.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pyfcm==1.4.5
PyJWT==1.6.4
pylint==2.1.1
pymongo==3.7.1
python-http-client==3.1.0
pytz==2018.5
requests==2.20.0
requests-toolbelt==0.8.0
sendgrid==5.6.0
six==1.11.0
sqlparse==0.2.4
typed-ast==1.1.0
urllib3==1.24
wrapt==1.10.11
